fix(scrollshot): improve Wayland scroll screenshot reliability
Fix occasional manual wheel unresponsiveness in Wayland scroll capture.
Optimize auto-scroll step and tolerate first frames without displacement.
Use millisecond timestamps to avoid inaccurate time-diff calculation.

修复 Wayland 环境下滚动截图手动滚轮偶发无响应的问题。
优化自动滚动步长，增加首帧无位移容错，降低误报概率。
调整滚动截图时间戳类型，避免时间差计算精度不足。

Log: 修复 Wayland 环境下滚动截图失败问题
Bug: https://pms.uniontech.com/bug-view-278311.html

+++ src/utils/scrollScreenshot.h
@@ -45,7 +45,7 @@ public:
     //手动滚动时的函数处理
     void setScrollModel(bool model); //设置滚动模式，先设置滚动模式，再添加图片
     QRect getInvalidArea();//获取调整区域
-    void setTimeAndCalculateTimeDiff(int time); //设置时间并计算时间差
+    void setTimeAndCalculateTimeDiff(qint64 time); //设置时间并计算时间差
 signals:
     void getOneImg();
     void updatePreviewImg(QImage img);
